refactor: IFS-2400 move Security and Berechtigungsmanager interfaces to core package
