+++ isy-security/src/main/java/de/bund/bva/isyfact/security/core/IsyOAuth2Security.java
@@ -6,8 +6,6 @@ import java.util.Set;
 import org.springframework.lang.Nullable;
 
 import de.bund.bva.isyfact.security.oauth2.client.Authentifizierungsmanager;
-import de.bund.bva.isyfact.security.Berechtigungsmanager;
-import de.bund.bva.isyfact.security.Security;
 import de.bund.bva.isyfact.security.xmlparser.RolePrivilegesMapper;
 
 /**
